requirements.txt
Django>=3.0.5
django-cors-headers == 3.*
django-extensions == 2.2.*
django-mysql == 3.4.*
djangorestframework == 3.11.*
requests == 2.23.*
uWSGI == 2.0.*
django-sendfile2 == 0.5.1
psycopg2-binary==2.8.3
mysqlclient == 1.4.*
six==1.14.*
python-dotenv == 0.13.*
markdown2 == 2.3.*
tokenlib == 2.0.*